chore: renamed and moved a util function
